Two item sale
A local department store is holding a sale! Buy one item, get second item 50% off.  If the customer is a card member with the store, an additional 10% is taken off the final price.  The user is queried for the price of two items, the truth value of being a card carrying member, and finally the local sales tax.  The program returns the price before any tax or discounts are applied, the price after applicable discounts, and then the price after applying the sales tax.