Phone call rates
Queries the user for the time that a call was initiated, in 24 hour format (HH:mm).  Queries the user next for the duration of the call in minutes, followed by the day of the week that the call was initiated.  The program returns the cost of the phone call as a dollar amount.  Phone calls are billed at different rates on nights and weekends.  The billing rate of the phone call is determined only by the time and day that the call was initiated, and does not prorate for calls that continue into different rate time ranges.  Night and weekend rates do not overlap.