Crappy roman numerals
User is queried for a positive integer.  The program returns as output a sequence of letters that is loosely representative of the roman numeral scheme.  The program does not follow all of the rules of the scheme, most apparent is concerning penultimate numbers.  Example: 4 will be displayed as IIII instead of IV, 9 is VIIII instead IX, and so on.