Yearly Calendar
User is prompted to provide a 4-digit integer representing a year, followed by a number 1 - 7 corresponding to the first day of that year, where 1 = Sunday, 2 = Monday... 7 = Saturday.  The program returns as output the entire calendar for that year.  The subroutine print_year_calender returns the day of the week of the next year's first day, in order support the program expansion of multiple year output.