Perfect and amicable numbers
User is queried for an integer ≥ 2.  Program returns all perfect numbers and amicable pairs between 2 and the entered integer.  A perfect number is any unsigned integer that is equal to the sum of its own divisors. Amicable number pairs are two positive integers where each number is equal to the sum of the other number's divisors.  Sidenote: amicable pairs required a lot of effort to solve.  TODO: Remove and relocate a global integer.