Word Counting
User is queried for a line of text.  The program returns as output words separated individually, one word per line.  Words are identified by ' ' space between text.  Non-erroneous input is a precondition.  Program does not sanitize input for profanity, use with caution. May be expanded or integrated with a file/database and behave as a rudimentary spellchecker or profanity censor.