Missing Numbers in a Range
User is queried for an amount of integers, n.  User is then prompted for entry of n integers in the range of 0 - n, inclusive.  For any possible user entry, there will always be at minimum 1 integer not entered by the user in the range 0 - n inclusive.  The program returns as output the integer(s) not selected by the user.