Number Search
User is queried for a sequence of positive integers until the user enters a negative integer.  User is then queried for an integer to search amongst the entered integers.  The program returns as output where searched integer may be found in the entered integers, if at all.  The program executes this functionality twice, although the code within is designed differently.