Exercise in Recursion
An exercise in recursive functions.  User is queried for a positive integer, n.  The returns as output 3 figures: a right triangle where the base and altitude equal n, a half-hourglass where the base equals n and the height equals 2n, and finally a "ruler" where n equals the maximum width of the ruler.