Jump-it
A simple, linear pathfinding algorithm.  The user is queried for a sequence of positive integers, starting with 0.  The program recursively determines the cheapest path to cross a numerical array, from arr[0] to arr[size -1].  The program will start at cell 0, and then determine to either move adjacently, or jump over the adjacent cell.  The program will recursively repeat this process until 3 or less cells remain.  The program returns as output the total, minimum cost to proceed across the array.