Classes: Money and checks
Included are two classes that test implement US currency as a derived data type.  A main driver function is included to demonstrate the functionality of the two classes. Does not utilize dynamic memory. Inheritance is not implemented.  Future improvements may see the separation of the classes and their functions into distinct files, the inclusion of more useful public functions, and additional cleanup with input sanitization.